Adding Identity Service Duende

+++ src/AuctionService/Controllers/AuctionController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -38,12 +39,15 @@ public class AuctionsController(AuctionDbContext _context, IMapper _mapper, IPub
 		return _mapper.Map<AuctionDto>(auction);
 	}
 
+	[Authorize]
 	[HttpPost]
 	public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
 	{
 		var auction = _mapper.Map<Auction>(auctionDto);
 
-		auction.Seller = "test";
+		// TODO: add current user as seller
+		auction.Seller = User.Identity.Name;
+
 		_context.Auctions.Add(auction);
 
 		var newAuction = _mapper.Map<AuctionDto>(auction); 
@@ -59,6 +63,7 @@ public class AuctionsController(AuctionDbContext _context, IMapper _mapper, IPub
 														newAuction );
 	}
 
+	[Authorize]
 	[HttpPut("{id}")]
 	public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
 	{
@@ -67,7 +72,7 @@ public class AuctionsController(AuctionDbContext _context, IMapper _mapper, IPub
 		
 		if (auction == null) return NotFound();
 
-		// TODO: check seller == username
+		if(auction.Seller != User.Identity.Name) return Forbid("");
 		
 		auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
 		auction.Item.Model = updateAuctionDto.Make ?? auction.Item.Make;
@@ -82,6 +87,7 @@ public class AuctionsController(AuctionDbContext _context, IMapper _mapper, IPub
 		return BadRequest("Problem saving changes");
 	}
 
+	[Authorize]
 	[HttpDelete("{id}")]
 	public async Task<ActionResult> DeleteAuction(Guid id)
 	{
@@ -89,11 +95,11 @@ public class AuctionsController(AuctionDbContext _context, IMapper _mapper, IPub
 
 		if (auction == null) return NotFound();
 
-		// TODO: check seller == username
+		if(auction.Seller != User.Identity.Name) return Forbid("");
 
 		_context.Auctions.Remove(auction);
 
-		await _publishEndPoint.Publish<AuctionDeleted>(_mapper.Map<AuctionUpdated>(new { Id= auction.Id.ToString()}));
+		await _publishEndPoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString() });
 
 		var result = await _context.SaveChangesAsync() > 0;
